Refactor attract_scores().
