+++ Makefile
@@ -6,11 +6,13 @@ CFG = nes.cfg
 LIB = runtime.lib
 
 SRCS = $(NAME).s \
-	attract_monsters.s
+	attract_monsters.s \
+	attract_scores.s
 
 OBJS = $(NAME).o \
 	crt0.o \
-	attract_monsters.o
+	attract_monsters.o \
+	attract_scores.o
 
 all: $(NAME).nes 
 
@@ -26,6 +28,12 @@ attract_monsters.o: attract_monsters.s
 
 attract_monsters.s: attract_monsters.c
 	$(CC65) -Oi attract_monsters.c --add-source
+
+attract_scores.o: attract_scores.s
+	$(CA65) attract_scores.s
+
+attract_scores.s: attract_scores.c
+	$(CC65) -Oi attract_scores.c --add-source
 
 $(NAME).o: $(NAME).s
 	$(CA65) $(NAME).s
