+++ attract_scores.h
@@ -1,0 +1,13 @@
+#ifndef ATTRACT_SCORES_H
+#define ATTRACT_SCORES_H
+
+#include "constants.h"
+#include "neslib.h"
+#include "wow_scores.h"
+
+/**
+ * attract_score() - show scores
+ */
+void attract_scores(void);
+
+#endif /* ATTRACT_SCORES_H */
