+++ wow.c
@@ -855,39 +855,6 @@ void run_dungeon(unsigned char dungeon_num)
 }
 
 /**
- * attract_score() - show scores
- */
-void attract_scores(void)
-{
-  vram_adr(NAMETABLE_C);
-  vram_fill(0,1024);
-
-  pal_bright(0);
-
-  // Dump scores into nametable A
-  vram_adr(NAMETABLE_A);
-  vram_unrle(wow_scores);
-  pal_bg(palette);
-  pal_spr(palette);
-  ppu_on_all();
-  ppu_wait_frame();
-  bank_spr(1);
-  bank_bg(0);
-  pal_fade_to(4);
-
-  while(1)
-    {
-      ppu_wait_frame();
-      ++frame_cnt;
-      if (frame_cnt==255) break;
-    }
-
-  pal_fade_to(0);
-  clear_update_buffer();
-  
-}
-
-/**
  * set_teleport(openclose)
  * openclose = 0 for open, 1 for close
  */
@@ -1152,7 +1119,7 @@ void main(void)
   
   while(1)
     {
-      /* attract_scores(); */
+      attract_scores();
       attract_monsters();
       /* run_dungeon(1); */
     }
